enabled asynchronous stopping of the current search refactoring to use more plain structs or unique_ptr

+++ engine/src/evalinfo.cpp
@@ -53,12 +53,12 @@ size_t EvalInfo::calculate_elapsed_time_ms() const
     return chrono::duration_cast<chrono::milliseconds>(end - start).count();
 }
 
-int EvalInfo::calculate_nps(size_t elapsedTimeMS) const
+size_t EvalInfo::calculate_nps(size_t elapsedTimeMS) const
 {
     return int(((nodes-nodesPreSearch) / (elapsedTimeMS / 1000.0f)) + 0.5f);
 }
 
-int EvalInfo::calculate_nps() const
+size_t EvalInfo::calculate_nps() const
 {
     return calculate_nps(calculate_elapsed_time_ms());
 }
